Improve method chaining model.(4th try; wip - 11)
