+++ src/main/java/com/github/dakusui/pcond/core/fluent4/sandbox/BooleanChecker.java
@@ -34,5 +34,15 @@ public interface BooleanChecker<T> extends Checker<
     public BooleanChecker<Boolean> rebase() {
       return new Impl<>(this::value, makeTrivial(Functions.identity()));
     }
+
+    @Override
+    public T baseValue() {
+      return this.baseValue.get();
+    }
+
+    @Override
+    public Function<T, Boolean> transformFunction() {
+      return this.transformFunction;
+    }
   }
 }
